+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/command_meta/delay/DelayInitializer.java
@@ -61,7 +61,9 @@ public class DelayInitializer extends ModuleInitializer {
     }
 
     private static void shutdownDelaySchedulerExecutor() {
-        DELAY_COMMAND_EXECUTOR.shutdown();
+        if (DELAY_COMMAND_EXECUTOR != null) {
+            DELAY_COMMAND_EXECUTOR.shutdown();
+        }
     }
 
     @Override