+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
-- [core] feature: enhance the exception handler for async tasks.
+- [core] feature: improve the exception handler for async tasks.
+- [core] feature: improve the exception handler for `/fuji reload` command.